Integrate newWallet and importWallet in create account and derive profile flows
